pyproject.toml
[project]
name = "v26meme"
version = "4.7.0"
description = "v26meme v4.7: Autonomous discovery + dynamic universe + PIT replay + paper execution."
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26", "pandas>=2.2", "pydantic>=2.7", "scipy>=1.11",
  "ccxt>=4.3.69", "plotly>=5.20", "streamlit>=1.35", "pyyaml>=6.0.1",
  "click>=8.1.7", "loguru>=0.7.2", "redis>=5.0", "pyarrow>=14.0",
  "scikit-learn>=1.3", "google-cloud-secret-manager>=2.16.0",
  "python-dotenv>=1.0.1", "requests>=2.31", "vaderSentiment>=3.3.2"
]
